feat: 파일명 수정 및 props,variant 정리
